improvements based on comments
